add reviewUsecases files for atomic db operations
