+++ src/usecases/reviewUsecases/deleteReviewFromPlace.js
@@ -1,0 +1,3 @@
+const deleteReviewFromPlace = async (id) => await Place.findByIdAndDelete(id)
+
+export { deleteReviewFromPlace }
