+++ src/usecases/reviewUsecases/deleteReviewFromRoute.js
@@ -1,0 +1,3 @@
+const deleteReviewFromRoute = async (id) => await Route.findByIdAndDelete(id)
+
+export { deleteReviewFromRoute }
