+++ src/usecases/reviewUsecases/getAllReviewsFromPlace.js
@@ -1,0 +1,3 @@
+const getAllReviewsFromPlace = async (id) => await Place.findById(id)
+
+export { getAllReviewsFromPlace }
