+++ src/usecases/reviewUsecases/getAllReviewsFromRoute.js
@@ -1,0 +1,3 @@
+const getAllReviewsFromRoute = async (id) => await Route.findById(id)
+
+export { getAllReviewsFromRoute }
