+++ src/usecases/reviewUsecases/postReviewToPlace.js
@@ -1,0 +1,3 @@
+const postReviewToPlace = async (id, review) => await Place.findById(id, review)
+
+export { postReviewToPlace }
