+++ src/usecases/reviewUsecases/postReviewToRoute.js
@@ -1,0 +1,3 @@
+const postReviewToRoute = async (id, review) => await Route.findById(id, review)
+
+export { postReviewToRoute }
