+++ src/usecases/reviewUsecases/updateReviewFromPlace.js
@@ -1,0 +1,4 @@
+const updateReviewFromPlace = async (id, body) =>
+  await Place.findByIdAndUpdate(id, body)
+
+export { updateReviewFromPlace }
