+++ src/usecases/reviewUsecases/updateReviewFromRoute.js
@@ -1,0 +1,4 @@
+const updateReviewFromRoute = async (id, body) =>
+  await Route.findByIdAndUpdate(id, body)
+
+export { updateReviewFromRoute }
